requirements.txt
feedparser>=6.0.0
requests>=2.31.0
openai>=1.0.0
python-dateutil>=2.8.0
pytz>=2024.1
